test/gthr_{cond,mutex}_timeout: `time_t` may be 32-bit sometimes

+++ test/gthr_mutex_timeout.c
@@ -18,11 +18,11 @@ main(void)
     int r;
 
     // Round the time up.
-    int64_t sleep_until = time(NULL) * 1000 + 2000;
+    int64_t sleep_until = int64_t() time(NULL) * 1000 + 2000;
     _MCF_sleep(&sleep_until);
 
     now = _MCF_perf_counter();
-    timeout.tv_sec = time(NULL) + 1;
+    timeout.tv_sec = int64_t() time(NULL) + 1;
     timeout.tv_nsec = 100000000;
     r = __gthread_mutex_timedlock(&mutex, &timeout);  // lock it
     assert(r == 0);
@@ -31,7 +31,7 @@ main(void)
     assert(delta <= 100);
 
     now = _MCF_perf_counter();
-    timeout.tv_sec = time(NULL) + 1;
+    timeout.tv_sec = int64_t() time(NULL) + 1;
     timeout.tv_nsec = 100000000;
     r = __gthread_mutex_timedlock(&mutex, &timeout);
     assert(r == ETIMEDOUT);
